requirements.txt
selenium~=4.26.1
pytest~=8.0.2
allure-pytest
pyyaml~=6.0.2
pytest-xdist
requests~=2.32.3
pytest-order
pytest-rerunfailures
faker